THX-34938 setup bintray publishing. clean up project dependencies by following dependencies.gradle pattern.  move to latest java plugin dependency configurations: implementation and testImplementation to avoid deprecation in gradle 5.  rename project files to align with new Thunderhead packaging.  Put publishing vars in gradle.properties.
