--- android-gradle-aspectj/build.gradle
@@ -1,93 +1,0 @@
-/*
- *    Copyright 2018 Thunderhead
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-buildscript {
-    ext.kotlin_version = '1.1.51'
-
-    repositories {
-        google()
-        jcenter()
-        mavenCentral()
-    }
-    dependencies {
-        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.7.3'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-    }
-}
-
-apply plugin: 'kotlin'
-apply plugin: 'maven'
-apply plugin: 'com.jfrog.bintray'
-
-project.ext {
-    aspectjVersion = '1.8.12'
-}
-
-dependencies {
-    repositories {
-        google()
-        jcenter()
-        mavenCentral()
-        maven { url 'https://maven.google.com' }
-        maven { url 'http://repository.jetbrains.com/utils' }
-        maven { url "https://dl.bintray.com/archinamon/maven" }
-    }
-
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version"
-    compileOnly gradleApi()
-    compileOnly 'com.android.tools.build:gradle:3.0.0'
-    compile "org.aspectj:aspectjrt:$aspectjVersion"
-    compile "org.aspectj:aspectjtools:$aspectjVersion"
-
-    testCompile 'junit:junit:4.10'
-    testCompile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    testCompile "org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version"
-    testCompile gradleApi()
-    testCompile 'com.android.tools.build:gradle:3.0.0'
-}
-
-group = 'com.archinamon'
-version = '3.2.0'
-
-task sourcesJar(type: Jar, dependsOn: classes) {
-    classifier = 'sources'
-    from sourceSets.main.allSource
-}
-
-artifacts {
-    archives sourcesJar
-}
-
-if (project.hasProperty('user') && project.hasProperty('apiKey')) {
-    bintray {
-        user = project.properties.user
-        key = project.properties.apiKey
-
-        configurations = ['archives']
-        pkg {
-            repo = "maven"
-            name = "android-gradle-aspectj"
-//            websiteUrl = siteUrl
-            vcsUrl = "https://github.com/Archinamon/GradleAspectJ-Android"
-            licenses = ["Apache-2.0"]
-            publish = true
-            version {
-                name = project.version
-                released = new Date()
-            }
-        }
-    }
-}
