--- android-gradle-aspectj/src/main/kotlin/com/archinamon/plugin/ConfigScope.kt
@@ -1,15 +1,0 @@
-package com.archinamon.plugin
-
-/**
- * TODO: Add description
- *
- * @author archinamon on 12/04/17.
- */
-
-internal enum class ConfigScope {
-
-    STANDARD,
-    PROVIDE,
-    EXTEND,
-    TEST;
-}
