--- android-gradle-aspectj/src/main/resources/META-INF/gradle-plugins/com.archinamon.aspectj.properties
@@ -1,1 +1,0 @@
-implementation-class=com.archinamon.plugin.AspectJWrapper$Standard
