+++ android-gradle-plugin-aspectj/src/main/kotlin/com/archinamon/api/transform/ProvidesTransformer.kt
@@ -14,5 +14,19 @@
  *    limitations under the License.
  */
 
-rootProject.name = 'ThunderheadAndroidGradleAspectJ'
-include ':android-gradle-plugin-aspectj'
+package com.archinamon.api.transform
+
+import com.android.build.api.transform.TransformInvocation
+import com.archinamon.utils.logBypassTransformation
+import org.gradle.api.Project
+
+internal class ProvidesTransformer(project: Project): AspectJTransform(project, BuildPolicy.SIMPLE) {
+
+    override fun prepareProject(): AspectJTransform {
+        return this
+    }
+
+    override fun transform(transformInvocation: TransformInvocation) {
+        logBypassTransformation()
+    }
+}
