+++ android-gradle-plugin-aspectj/src/main/kotlin/com/archinamon/plugin/ConfigScope.kt
@@ -14,5 +14,18 @@
  *    limitations under the License.
  */
 
-rootProject.name = 'ThunderheadAndroidGradleAspectJ'
-include ':android-gradle-plugin-aspectj'
+package com.archinamon.plugin
+
+/**
+ * TODO: Add description
+ *
+ * @author archinamon on 12/04/17.
+ */
+
+internal enum class ConfigScope {
+
+    STANDARD,
+    PROVIDE,
+    EXTEND,
+    TEST;
+}
