+++ android-gradle-plugin-aspectj/src/test/kotlin/com/archinamon/utils/AarExploringTest.kt
@@ -14,5 +14,21 @@
  *    limitations under the License.
  */
 
-rootProject.name = 'ThunderheadAndroidGradleAspectJ'
-include ':android-gradle-plugin-aspectj'
+package com.archinamon.utils
+
+import org.junit.Rule
+import org.junit.Test
+import org.junit.rules.TemporaryFolder
+import java.io.File
+
+class AarExploringTest {
+
+    @Rule
+    @JvmField
+    val folder = TemporaryFolder()
+
+    @Test
+    fun noNPEIfFileNotExist() {
+        findPackageNameIfAar(File("/build-cache/this/file/not/exist"))
+    }
+}
