+++ gradle.properties
@@ -1,3 +1,19 @@
+#
+#    Copyright 2018 Thunderhead
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+#
+
 # Project-wide Gradle settings.
 
 # IDE (e.g. Android Studio) users:
@@ -15,3 +31,21 @@
 # When configured, Gradle will run in incubating parallel mode.
 # This option should only be used with decoupled projects. More details, visit
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
+pluginVersion=4.0.0
+pluginName=android-gradle-plugin-aspectj
+pluginDescription=A Gradle plug-in adding support for AspectJ in an Android project.
+publishedGroupId=com.thunderhead.android
+artifact=android-gradle-plugin-aspectj
+bintrayRepo=Android
+bintrayName=android-gradle-plugin-aspectj
+bintrayUserOrg=thunderheadone
+siteUrl=https://www.thunderhead.com
+gitUrl=https://github.com/thunderheadone/one-android-gradle-aspectj.git
+githubRepository=thunderheadone/one-android-gradle-aspectj
+issuesUrl=https://github.com/thunderheadone/one-android-gradle-aspectj/issues
+developerId=Thunderhead
+developerName=Thunderhead
+developerEmail=onesupport@thunderhead.com
+licenseName=The Apache Software License, Version 2.0
+licenseUrl=http://www.apache.org/licenses/LICENSE-2.0.txt
+allLicenses=["Apache-2.0"]
