+++ gradleScripts/dependencies.gradle
@@ -1,0 +1,41 @@
+/*
+ *    Copyright 2018 Thunderhead
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+ext {
+
+    kotlin_version = '1.1.51'
+    aspectj_version = '1.8.12'
+    junit_version = '4.10'
+    gradlePlugins = [
+            kotlin: "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version",
+            bintray: 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.7.3'
+    ]
+    jetbrains = [
+            kotlin_std_lib_java_8: "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version",
+            kotlin_std_lib: "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version",
+            kotlin_test_junit: "org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version"
+    ]
+    aspectj = [
+            runtime: "org.aspectj:aspectjrt:$aspectj_version",
+            tools: "org.aspectj:aspectjtools:$aspectj_version"
+    ]
+    testLibraries = [
+            junit: "junit:junit:$junit_version"
+    ]
+    android = [
+            buildTools: 'com.android.tools.build:gradle:3.0.0'
+    ]
+}
